% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabnew.r
\docType{data}
\name{maize}
\alias{maize}
\title{Maize yield trial data}
\format{
An object of class \code{data.frame} with 260 rows and 4 columns.
}
\usage{
maize
}
\description{
Maize yield trial data
}
\examples{
str(maize)
}
\references{
Fan X.M., Kang M.S., Chen H.M., Zhang Y.D., Tan J., Xu C.X. (2007) Yield stability of maize hybrids evaluated in multi-environment trials in Yunnan, China. Agronomy Journal.99:220-228
}
\keyword{datasets}
